requirements.txt
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
Beaker==1.10.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
idna==2.6
ipdb==0.11
ipython==5.7.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
MetPy==0.8.0
numpy==1.14.4
pandas==0.23.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
Pint==0.8.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scandir==1.7
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
xarray==0.10.7
